Shipment of deepseek-v2-lite-base on autocomplete for PLG users (#5079)
## Context
1. This PR adds a feature flag `cody-autocomplete-deepseek-v2-lite-base`
to ship the deepseek-v2-lite-base model. For now, the model will be
shipped for all the PLG users.
2. Sends a `anonymousUserId` fields in the fireworks payload. This
payload is send to fireworks so that each user can be redirected to the
same deployment and can leverage prompt caching.

## Test plan
1. Override myself in the feature flag and ensure we are hitting
`deepseek` model when on dotCom instance.
2. Override myself in the feature flag and ensure we are hitting
`starcoder` model when on S2 instance.
3. Checked that we pass `anonymousUserId` correctly to CG.

## Changelog
1. Shipping `deepseek-coder-v2-lite-base` as the default autocomplete
model for PLG users.
2. Add `anonymousUserId` in the CG payload to leverage fireworks prompt
caching.