Agent: add new sidebar panel agent protocol & webview sign-in form (#5192)
Part of
https://linear.app/sourcegraph/issue/CODY-3208/user-friendly-login-for-plg-and-enterprise
https://linear.app/sourcegraph/issue/CODY-3055/web-based-login-ui-may-spuriously-appear

- Add a new `chat/sidebar/new` request to the agent protocol that
creates a new chat panel with a sidebar webview type
- Update the `vscode-shim` to notify the agent when the window context
changes, e.g. when the user's authentication status changes
- Update the `ChatController` to handle `signin` and `signout` messages
from the webview
- Add a new `ClientSignInForm` component for clients that don't support
the sign-in quickpick

These changes address the following issues:
- 'chat/sidebar/new' enables clients to start a new sidebar chat. The
current 'chat/new' starts an editor chat where the nav bar is removed
from the chat view
- Clicking on the sign-out button on the client Account page will sign
users out and send a `window/didChangeContext` notification to clients,
this allows JetBrains and other clients to display the login page / chat
view based on the value using by the webview. Part of
https://linear.app/sourcegraph/issue/CODY-3055/web-based-login-ui-may-spuriously-appear

NOTE: Reason I didn't create a new field for client capability is
because we will eventually move into a unified authentication page,
which is currently in-design. The sign-in form in this is just a
temporary workaround for clients that do not have access to quick pick
or on sign out.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

1. cd into vscode directory and run `pnpm storybook`
2. adjust the `codyIDE` value to confirm the current Login form for VS
Code and Web remains unchanged
3. Verified the Login form only shows up for codyIDE that is not VS Code
and Web


![image](https://github.com/user-attachments/assets/73b33361-c2f5-42c6-add2-b54cb34fbd9f)

codyIDE = VisualStudio / JetBrains / Eclipse

![image](https://github.com/user-attachments/assets/6e36c642-441e-48fc-829d-b092d855a7c0)

codyIDE = VSCode

![image](https://github.com/user-attachments/assets/179a472e-bf45-40e0-a270-0ee5317aa102)


## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->