fixed(web): removed `fs-extra` (#5194)
There's a current bug somewhere (I think in Vite?) that tries to
externalize node dependencies, and print out a warning message. But when
it loads the dependencies of `fs-extra`, specifically `graceful-fs` it
comes in as a symbol instead of string, which can't be interpolated and
it crashes.

![image](https://github.com/user-attachments/assets/b1c78ea2-0c59-4657-8d47-2ae4e03987e1)

Our only use of it is to copy a file, which we can do with `fs` so it is
unneeded.
## Test plan
Compiles and runs in browser.