feat(chat): add ChatEnvironmentContext to track client type
This change adds a new `ChatEnvironmentContext` to the Cody webview application. The context provides information about the client type (e.g. VSCode, Sourcegraph) which is used to customize the behavior of certain UI components, such as the `MarkdownFromCody` component.

The key changes include:

- Add `ChatEnvironmentContext` and `ChatEnvironmentContextData` types
- Provide the client type in the `chatEnvironmentContext` value in the `App` component
- Pass the `chatEnvironmentContext` to the `getAppWrappers` function and wrap it in the context provider
- Use the `clientType` value from the `ChatEnvironmentContext` in the `MarkdownFromCody` component to select the appropriate URL transformation logic

This change will allow the Cody webview to better adapt to the environment it is running in, improving the overall user experience.