do not clobber other @-mentions when initial context changes (#5121)
If you manually @-mention a file in Cody chat and then change your
current editor selection, your manual @-mention is clobbered. This was a
bug introduced when we changed the @-mention nodes from Lexical
TextNodes to DecoratorNode-based nodes. Now we need to traverse the tree
and check all nodes instead of just checking the TextNodes.



## Test plan

As described above.

## Changelog

- Fix a chat issue where manually entered @-mentions would be clobbered
if you changed your selection or active file in the editor.