fix a bug where the current file would not show up when searching for @-file mentions
This was unintentionally caused by https://github.com/sourcegraph/cody/pull/5057. The intent was to not show 2 mention items `Current File: foo.ts` and `foo.ts`, but it meant that if you typed `@foo` then neither would show up.

Also simplify how initial context is passed to MentionMenu. Just include it in `items` so we don't spread responsibility for displaying and filtering it across multiple components.