added/agent: persistent global state for agent (#5133)
Adds a var to ClientCapabilities allowing extensions to specify a file
path where the agent can persist global state, such as chat. Currently
does the simplest possible thing which is to write the entire state out
to disk on every update. This can obviously be improved, but at the cost
of complexity.

## Test plan
Added unit tests
Manually tested this with the Eclipse plugin in
https://github.com/sourcegraph/eclipse/pull/39

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog
Added ability to persist state from the agent

---------

Co-authored-by: Beatrix <beatrix@sourcegraph.com>
Co-authored-by: Olafur Geirsson <olafurpg@gmail.com>