feat: add ability to export chat history
Update the export chat history button to allow users to export their chat history as a JSON file on click using the native browser dialog. The functionality is still the same, but this change makes the feature available to all Cody clients by removing the use of VS Code API for save dialog.

The implementation involves:
- Adding a new `onDownloadChatClick` prop to the `TabsBar` component that is responsible for generating and downloading the chat history JSON file.
- Updating the `HistoryTab` component to include the new "Export History" button that calls the `onDownloadChatClick` callback.
- Ensuring the "Export History" button only requies VSCode users for confirming deletion, as the VS Code API might not be avaliable to all clients that is required for clearing the chat history.

Notes: The `TabsBar` component now requires the `onDownloadChatClick` prop to be provided.