fix more @-mention menu glitchiness by fixing height (#5224)
See code comments. Now, the orientation won't flip, and sometimes it
flipped the wrong way if it was 0px tall for a brief period (and then
did not re-detect the available height after it filled in with content).

## Test plan

In the Chat storybook, type `@` and ensure the menu is not obscured.

In the web demo or the Transcript storybook, type `@a` and then delete
so you just have `@`. The menu should be positioned correctly.

## Changelog

Fixed glitches in the positioning of the @-mention menu.