lazily import js-tiktoken to cut ~6mb from initial bundle size (#5231)
`js-tiktoken` is a very large JavaScript library. This changes how we
use it, cutting ~6mb from the initial bundle size (3mb for each of the
webviews and agent).

- Now, we only import `cl100k_base` and not the other tokenizers. These
were being tree-shaken anyway, but this protects us from regressions
there.
- We now lazily import the `cl100k_base` data, which means all token
counting functions are now async. In the webviews and on the web, this
results in a significant speedup in page load (~6mb smaller initial
bundle and ~25% faster DOMContentLoaded).



## Test plan

n/a