use the full width of the Cody chat viewport (#5237)
Many people have asked for Cody chat to use the full width in the
viewport. This is especially important now that we often show chat in a
narrow sidebar instead of an editor panel.

Example request: https://x.com/riipandi/status/1825414693860856206 (this
was just now, there have been many others)

This is a partial impl of this Figma:
https://www.figma.com/design/f078wFMKsIOaEwj7Iwj5xy/%E2%AC%9C%EF%B8%8F--Unified-Cody?node-id=4157-66156&t=UY3VI8o9wD53jilb-0.


![image](https://github.com/user-attachments/assets/0b60b5f9-1c3d-47b4-a7cb-23e51ba2443b)


## Test plan

Use Cody chat on a narrow screen.

## Changelog

The chat UI now uses the full width of its viewport, which is helpful
when using chat in a narrow editor sidebar.