.gitignore `protocol_generated` and `copyProtocol` before `kotlinCompile`  (#2072)
The former approach allowed us to have a protocol that does not
correspond to the commit pointed by `cody.commit`. Usually, we had an
outdated protocol. `cody.commit` corresponds to the specific protocol
and it seems to be a good idea to keep the protocol strictly up to date.

This PR removes the generated protocol from git and runs `copyProtocol`
just before `kotlinCompile` task. It guarantees us the consistency. It
allows us to notice some important changes earlier (in test CI).


## Test plan
1. Green CI