Chat: update FileLink component to work with Agent (#5260)
The main goal of this PR is to enable FileLink component to work across
clients. Currently, the FileLink is using the "commandURL" that's only
supported inside VS Code, which means all the other clients will need to
implement additional logic to handle opening a link in file. This PR
removes the use of "command URL" and replace the links with button that
invoke the `vscode.open` API that is available in Agent to open file
links, instead of the embedded command links.

This PR also includes the following change:
- Adds support for the Eclipse IDE in the Cody application. The
`AgentWorkspaceConfiguration` class has been updated to handle the new
`'eclipse'` case, and the `CodyIDE` enum in the shared configuration has
been extended to include the `Eclipse` value.
- Add `multipleWebviewsEnabled` to webview config based on the client
capability and use the new value to determine if we should render editor
webview type or not.
- The `MarkdownFromCody` component has been updated to include the
`Eclipse` IDE in the URL processor mapping, and the `TabsBar` component
has been modified to conditionally render the "Open in Editor" and
"Settings" tabs based on the `multipleWebviewsEnabled` configuration.
-`CodyWebPanel` component has been updated to set the
`multipleWebviewsEnabled` configuration to `false` for the web-based
Cody application.

## Test plan

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

All the current E2E tests that cover FileLinks have been updated to use
the button element instead of link element. Functionality was not
changed.

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->