Chat: default to sidebar location for chat instead of sticky
Previously, we used the setting `'sticky'` by default to determine the
location of the chat UI. This was problematic because the implementation
of `'sticky'` can feel buggy sometimes, it was easy to reproduce
scenarios where triggering alt+l in a row would cycle between the
sidebar, editor, and code. This PR works around this issue by defaulting
to the sidebar instead, making the alt+l shortcut have more consistent
behavior.

We can discuss changing the default to `'sticky'` again when the
implementation has more predictable behavior.