Revert: rehype-highlight bump (#5313)
There has been a (long ongoing) issue with `rehype-highlight` /
`highight-js` causing OOM issues on `>v6.0.0`. This reverts a recent
version bump that might have [re-introduced this
issue](https://github.com/sourcegraph/cody/pull/4459) in Cody chat
causing grey-screens and crashes.

> Important: I wasn't able to re-produce the issue myself for some
reason. So this is a bit of a "shot-in-the-dark" fix. Hopefully we can
have those who reported the issue validate if it actually worked.
Lowering the version shouldn't hurt any other changes though as it was
merely done to reduce the bundle size somewhat.

## Test plan
CI