Smart Apply: support writing to empty files
Fixes CODY-3421

Previously, smart apply didn't work for empty files, it would create a
new file in the root directory instead. This was problematic because
users could create an empty file in a subdirectory, ask chat to write
code, press smart "Apply" and Cody would open a different file with the
same name (but in the root directory).

This PR fixes the problem by adding empty files to the context if those
files are explicitly added as context items in the chat input.
Previously, we skipped empty files, presumably because they add limited
value when pulled implicitly (via embeddings or keyword search).