Guardrails: fix DDoS bug
Fixes SRCH-942

Previously, if an enterprise user had a chat window open with a code block
containing more than 10 lines of code, then Cody would send a request to
our Guardrails server on every single editor event. This was caused by a
`useEffect` block that has a large number of dependencies where one of
the dependencies (`humanMessage`) was not memoized.

This PR fixes the bug by wrapping `humanMessage` in a `useMemo` block.
Ideally, the code should be refactored a bit to have more narrow
`useEffect` blocks with fewer dependencies.