added(agent): `chat/import` to import historical chats (#5304)
Adds a new `chat/import` request that allows editors that had custom
chat implementations (JetBrains) to import their chat history to the
agent so that it can be used in the webview.

## Test plan
Manually tested with JetBrains. Need to add some agent tests.

## Changelog
Added ability to import historical chats through the agent.