fix storybook CSS
The PR https://github.com/sourcegraph/cody/pull/5344 made it so that Tailwind CSS was not being emitted, which broke CSS styles in the storybook.

It should not be necessary to specify the postcssNested plugin here, as the referenced PostCSS config file already references it. This reduces duplication of PostCSS config and makes it so that storybooks' Tailwind CSS is emitted.