Chat: flip order of initial context chips
Previously, the chat input defaulted to using the chips REPO_CONTEXT
FILE_CONTEXT. This PR flips the order so that REPO_CONTEXT appears after
FILE_CONTEXT. The motivation for this change is to make it easier for
users to disable repo context (press backspace twice). The reasoning
is that users more frequently disable repo context than file context.