fix/JetBrains: Don't open a new chat window when applying code from chat (#5358)
Fixes CODY-3519.

This was spawning a new chat window because of an old code path that
assumed reusing chat webviews was futile in Agent. But now some clients
have native webviews, we can reuse them.

## Test plan

Tested manually in JetBrains:

1. Move the sidebar chat into the editor.
2. Split the editor right, so you have a file on one hand and a panel
chat on the other.
3. In the chat, @-tag the file and suggest some edits.
4. Hit Apply.
5. Accept the code changes.

Verify that this does NOT open a new chat panel.