Chat: fix bug where the "new chat" could crash the agent (#5363)
Fixes CODY-3510

While testing the Eclipse plugin, I was able to get the webview into a
permanently broken state with the reproduction:

- Send message `@Hello.java explain`
- Press "New Chat" while it's streaming the reply

The root cause was that we threw an exception when trying to store the
chat history with an error message. This PR works around the issue by
not storing messages with errors. However, I'm not 100% satisfied with
the fix because

- The process shouldn't permanently error if an exception is thrown
- I wasn't able to reproduce this in VS Code, only Eclipse and
JetBrains. This indicates we're not 100% reproducing the VCS
environment.

Either way, this should be a low-risk fix to merge.


## Test plan

Run reproduction steps in Eclipse.

<!-- Required. See
https://docs-legacy.sourcegraph.com/dev/background-information/testing_principles.
-->

## Changelog

<!-- OPTIONAL; info at
https://www.notion.so/sourcegraph/Writing-a-changelog-entry-dd997f411d524caabf0d8d38a24a878c
-->