Remove CodebaseStatusProvider (#5243)
Removes `CodebaseStatusProvider`, which previously was responsible for
monitoring the availability of enhanced context for each codebase that
was available. The enhanced context UI no longer exists, which means
this class is no longer necessary (and indeed was no longer used in most
places). Removing it allows us to remove the overhead imposed on local
context providers to provide the status of indexes.

The new @-mention-based UI has as different way of determining the
default repository chip. Currently it does not appear that it properly
handles multiple workspace folders. I.e., if I have 2 workspace folders
open, the @-mention completer does not allow me to tag the second
folder. We should fix this, but this is a current issue, so this PR does
not regress behavior. Users can also tag as many remote repositories
when connected to an enterprise instance.

We add a `WorkspaceReposMonitor` class that's responsible for observing
changes to workspace folders and keeping track of whether the
repositories in the current workspace are public. This is used only by
telemetry logic. See the tests in
`vscode/src/repository/repo-metadata-from-git-api.test.ts` for the
behavior (we err on the side of being conservative).

Removal of `CodebaseStatusProvider` allows us to remove a lot of dead
and rather complex code, which should help significantly with unifying
the context logic on the client side (this was the original motivation
for doing this).