added(telemetry): separate field for IDE name and server identification (#5276)
Pre 5.6 versions of Sourcegraph would reject requests from unknown
clients. This was resolved in
https://github.com/sourcegraph/sourcegraph/pull/63855, but for
enterprise customers who have not updated, we still need to identify
ourselves as JetBrains. This meant that all telemetry events would
falsely identify themselves as other JetBrains.

This introduces a field we hope to delete that allows a client to
identify as itself, EXCEPT for with the server. Once everyone is up to
5.6, we can remove this field.

Additionally, we check if the client is stateful for login events, so we
can report reliable installed and login events for agent users.

## Test plan
Tested in Eclipse

## Changelog
Allows identifying as Eclipse (or other previously unrecognized IDEs) in
telemetry events.