simplify initial AuthProvider status state
Previously, the initial state had the `AuthState.endpoint` value set to the string `'init'`. This was a special sentinel value that was not documented using the type system. Now, the value in the `AuthProvider` member field is `null` until it has been initialized.

Also renames `AuthProvider.getAuthState` to `AuthProvider.state`.