Fixes the heading of cmk cmdk component
In vscode cmdk component is using the input background. Input backgrounds are different from another background in vs code land.

On JetBrains, its not feasible to use input background as the background of these headers as they will appear with the same background as the panel, making the header lose its visual relevancy