clean up AuthStatus and AuthProvider (#5420)
No behavior change, just code cleanups.

### compute AuthStatus.isDotCom instead of denormalizing as prop

It was not kept consistent with AuthStatus.endpoint, which caused
confusion and added complexity. For example:

- Initially, `endpoint == 'init'` and `isDotCom == true`.
- When the user switches from a non-dotcom account to a dotcom account
or vice versa, the `isDotCom` prop is not updated until the auth
successfully completes.
- In many tests, `endpoint == ''` and `isDotCom == true`.

This is type-lying, and this change eliminates it.

### simplify AuthProvider, extract helpers to separate funcs

Previously, AuthProvider had a mix of state ownership, state transition
logic, and UI logic. Now, the UI logic is extracted to separate
functions. This is not the ideal permanent situation, but it makes it
easier to refactor AuthProvider's state handling in isolation.


### simplify initial AuthProvider status state

Previously, the initial state had the `AuthState.endpoint` value set to
the string `'init'`. This was a special sentinel value that was not
documented using the type system. Now, the value in the `AuthProvider`
member field is `null` until it has been initialized.

Also renames `AuthProvider.getAuthState` to `AuthProvider.state`.

## Test plan

CI