fixed(chat): Added a migration to local storage to update chat histor… (#5430)
Cherry pick of the merge commit from
https://github.com/sourcegraph/cody/pull/5407

## Changes

There was a bug in https://github.com/sourcegraph/jetbrains/pull/2108
where we imported chat's with ID's but the app actually expected them to
be UTC date strings. This caused some errors where it tried to do time
grouping on the UUID's and if those conversations were continued, they
would actually lose all of their timestamps.


![image](https://github.com/user-attachments/assets/d900a45c-f21c-4c6b-9b57-58e3a5b45d46)

This adds a migration that will be run once at startup to try to detect
and patch any corrupted conversations. Additionally this PR should fix
the migration for newer users.

## Test plan
Unit Tests and manually ran in JB

## Changelog
Fixed a bug in continuing conversations from JB in the webview that
caused them to be displayed as NaN weeks ago

Co-authored-by: James McNamara <jamesscottmcnamara@gmail.com>