simplify the "authed but Cody is not enabled" state (#5437)
Previously, we had 2 fields on AuthState: `authenticated` and
`isLoggedIn`. These are synonyms but actually meant 2 different things:

- `authenticated`: whether the user is authenticated
- `isLoggedIn`: whether the user is authenticated AND Cody is enabled on
the instance AND the user has a verified email if the instance requires
such

Neither of the additional constraints for `isLoggedIn` were handled
cleanly in the UI. So, this change makes it so that users are prevented
from signing in if Cody is not enabled (and are shown an error message
during sign-in). The verified email issue is a much lesser issue now (it
was an issue when we had different signup methods ~11 months ago) and is
handled by the server returning an error. In general, we should
implement these checks on the server not the client since the auth flow
is on the server already (and that is how similar things would be
implemented today now that we have the nice non-raw-access-token flow).

## Test plan

Set `{"cody.enabled": false}` in site config and try signing in. Confirm
that an error message is shown.