Chat: make "insert code at cursor" replace selection
Previously, the "Insert code at cursor" button always added the code
at the start of the cursor or text selection. This was problematic
because when you have a text selection, you intuitively expect it to
replace the whole selection, not add the code. This PR fixes the issue
by replacing when there is a text selection, and adding when there is a
cursor.