fix: handle git diff path parsing on Windows
This commit fixes an issue where the `Generate Commit Message` command would fail on Windows due to incorrect parsing of the git diff output. The change ensures that the file path is correctly identified even when the diff output does not contain the full file path.

CLOSE
https://linear.app/sourcegraph/issue/CODY-3601/bug-error-failed-to-get-git-output-when-using-commit-generation-on

## Test plan

Manually tested the `Generate Commit Message` command on both Windows and Mac to ensure it works as expected.

## Changelog

- Command: Fixed an issue where the experimental `Generate Commit Message` command would fail on Windows due to incorrect parsing of the git diff output.