Dev: fix params used by groqChatClient
Fix: https://linear.app/sourcegraph/issue/CODY-3672

This PR fixes an issue where the `chatParams` was being passed as a parameter object when using the local GROQ model provider, causing all network requests to fail when using the local config models.

The fix is to properly destructure the `chatParams` object before passing it to the fetch request.

## Changes

- Destructure the `chatParams` object before passing it to the fetch request in the `groqChatClient` function.

## Test Plan
- Verified that the local GROQ model provider now works correctly when using the `chatParams` object.

Changelog: Fixed an issue where the `chatParams` was being passed as a parameter object when using local model provider that caused some requests to fail. The chatParams are not probably deconstructed instead getting passed as an object. [pull/5483](https://github.com/sourcegraph/cody/pull/5483)