react to config/feature-flag changes using Observables in autocomplete (#5478)
Make it so our autocomplete logic reacts to config and feature flag
changes using Observables, instead of the current situation where there
are many different ad-hoc ways configuration and feature flag values are
passed through. This simplifies (or at least standardizes) the code,
reduces bugs, and (once this kind of change is applied everywhere in our
code) eliminates the need for users to reload the editor to pick up new
config.

## Test plan

e2e tests. Also try changing Cody autocomplete configs and signing
in/out and ensure that the right config is loaded.