Chat: shorten remote repository titles in mention menu
CLOSE https://linear.app/sourcegraph/issue/CODY-3499/increase-width-of-remote-repo-selector-in-vs-code-sidebar

The repository titles in the mention menu were displaying the full repository path, which could make the menu items too long to be displayed in the dropdown menu.

This change shortens the repository titles to only show the repository name (the part after the last slash) rather than including the codebase as display name since this open more spaces to display the actual repository name, and secondly, it is not common for our customers to host the same repository in two different codehost, and then have both of them imported to their Sourcegraph instance.