make some services more reactive to config
Partial application of https://github.com/sourcegraph/cody/pull/5221 to some services: FeatureFlagProvider, ClientStateBroadcaster, UpstreamHealthProvider, OpenTelemetryService, SentryService, etc.

The idea is to make services listen to config instead of checking it once and not updating themselves when config changes, which is bad because users need to reload the editor sometimes when they change config.

Co-authored-by: Valery Bugakov <skymk1@gmail.com>