API: move to api-version=2 for faster streaming LLM responses
Previously, the Cody IDE client used `api-version=1` for LLM chat
responses, which sent the full LLM response on every streaming change.
This was problematic because it meant we were sending a lot of redundant
traffic between the IDE client and the remote Sourcegraph instance
eventually resulting in a slower user experience.

Now, we use `api-version=2`, which only sends the delta between chunks.
This significantly reduces the amount of characters we're sending over
the wire. For example, for a chat response with 3k output tokens, we
were previously processing up to 1.8m tokens(!!) while now we only
process 3k tokens.